LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-remove-source-expected.txt


FAIL pointer updates (removing source elements) assert_equals: error events on c expected 1 but got 0
